Factor out plan formatter from test suite to main src. #309
