+++ compiler/src/test/scala/ingraph/sandbox/CompilerTest.scala
@@ -4,6 +4,7 @@ import ingraph.compiler.CypherToQPlan
 import ingraph.compiler.cypher2qplan.CypherParser
 import ingraph.compiler.exceptions.{CompilerConfigurationException, IncompleteCompilationException, IncompleteResolutionException}
 import ingraph.compiler.qplan2jplan.{JPlanToFPlan, QPlanToJPlan}
+import ingraph.compiler.util.FormatterUtil
 import ingraph.emf.util.PrettyPrinter
 import ingraph.model.expr.{EStub, ResolvableName}
 import ingraph.model.fplan.{FNode, LeafFNode}
@@ -33,7 +34,10 @@ case class CompilerTestConfig(querySuitePath: Option[String] = None
 
 abstract class CompilerTest extends FunSuite {
   val config = CompilerTestConfig()
-  val separatorLength = 77
+  val separatorLength = FormatterUtil.separatorLength
+  val printlnSuppressIfIngraph = FormatterUtil.printlnSuppressIfIngraph
+  def formatPlan(stuff: Any, heading: Option[String] = None, out: String => Unit = printlnSuppressIfIngraph): Unit = FormatterUtil.formatPlan(stuff, heading, out)
+
 
   def constructQueryFilePath(fileBaseName: String): String = {
     if (config.querySuitePath.isEmpty) {
@@ -69,26 +73,26 @@ abstract class CompilerTest extends FunSuite {
         printlnSuppressIfIngraph("-" * separatorLength)
       }
     }
-    if (config.printQuery ) formatStuff(query, Some("Query"))
+    if (config.printQuery ) formatPlan(query, Some("Query"))
 
     val cypher = CypherParser.parseString(query)
-    if (config.printCypher) formatStuff(PrettyPrinter.format(cypher), Some("Parsed query"))
+    if (config.printCypher) formatPlan(PrettyPrinter.format(cypher), Some("Parsed query"))
 
     val qplan = CypherToQPlan.build_IKnowWhatImDoing(cypher
       , queryName
       , skipResolve = config.skipQPlanResolve
       , skipBeautify = config.skipQPlanBeautify
     )
-    if (config.printQPlan ) formatStuff(qplan)
+    if (config.printQPlan ) formatPlan(qplan)
 
     assertNoStub(qplan)
     if (!config.skipQPlanResolve) assertResolved(qplan)
 
     val jplan = if (config.compileQPlanOnly) null else QPlanToJPlan.transform(qplan)
-    if (config.printJPlan ) formatStuff(jplan)
+    if (config.printJPlan ) formatPlan(jplan)
 
     val fplan = if (config.compileQPlanOnly) null else JPlanToFPlan.transform(jplan)
-    if (config.printFPlan ) formatStuff(fplan)
+    if (config.printFPlan ) formatPlan(fplan)
 
     return CompilationStages(qplan, jplan, fplan)
   }
@@ -96,28 +100,6 @@ abstract class CompilerTest extends FunSuite {
   def getLeafNodes(plan: FNode): Seq[FNode] = {
     if (plan.isInstanceOf[LeafFNode]) return plan :: Nil
     return plan.children.flatMap(x => getLeafNodes(x))
-  }
-
-  /**
-    * Formats a {Q,J,F}Plan, or virtually anything and send to the out channel
-    * @param stuff {Q,J,F}Plan instance, or any other that has toString. In sace null is passed, nothing will be sent to the out channel.
-    * @param heading The heading line for the formatted plan. In case it was a {Q,J,F}Plan, this heading is inferred if omitted, otherwise this must be supplied.
-    * @param out The out channel method, defaults to suppressing output if INGRAPH_COMPILER_TEST_SUPPRESS_PRINTLN environment variable is defined, and println otherwise
-    */
-  def formatStuff(stuff: Any, heading: Option[String] = None, out: String => Unit = printlnSuppressIfIngraph): Unit = {
-    val _heading: String = heading.getOrElse(
-      stuff match {
-        case _: QNode => "QPlan"
-        case _: JNode => "JPlan"
-        case _: FNode => "FPlan"
-        case null => return
-      }
-    )
-
-    out(s"${_heading}:")
-    out("-" * (_heading.length + 1))
-    out(stuff.toString)
-    out("-" * separatorLength)
   }
 
   def assertNoStub(q: QNode): Unit = {
@@ -143,14 +125,6 @@ abstract class CompilerTest extends FunSuite {
       case _ => false
     }, itn).fold[Unit](Unit)( e => throw new IncompleteResolutionException(e.toString))
   }
-
-  /**
-    * println is returned unless condition holds.
-    */
-  def printlnSuppressIf(condition: Boolean): String => Unit =
-    if (condition) _ => {} else Predef.println
-
-  def printlnSuppressIfIngraph = printlnSuppressIf(sys.env.get("INGRAPH_COMPILER_TEST_SUPPRESS_PRINTLN").isDefined)
 
   /** Override println hack. Use printlnSuppressIfIngraph() instead") */
   @Deprecated
